project/Code/requirements.txt
langchain
faiss-cpu
openai
pypdf
sentence-transformers
